% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n1pas_method.R
\name{get_OR}
\alias{get_OR}
\title{Computes odds ratios for each pathway to quantify enrichment of alternatively spliced genes.}
\usage{
get_OR(tmp_genes, dist_data, genes_range = c(15, 500))
}
\arguments{
\item{tmp_genes}{a character vector specifying the gene symbols annotated to the pathway of interest.}

\item{dist_data}{a data frame indicating whether a gene is alternatively spliced or not.}

\item{genes_range}{a two-component vector to filter gene sets based on minimum and maximum number of genes. Default is (15, 500).}
}
\value{
a list containing the odds ratio and number of measured genes in the pathway.
}
\description{
\code{get_OR} computes odds ratios for each pathway to quantify enrichment of alternatively spliced genes.
}
\seealso{
\code{\link{compute_hellinger}} \code{\link{get_OR}} \code{link{transform_iso_gene}} \code{link{transform_gene_pathway}} \code{\link{blca_patient_iso_kegg}}
}
