% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n1pas.R
\docType{package}
\name{n1pas}
\alias{n1pas}
\alias{n1pas-package}
\title{\code{n1pas}: An R Package for N-of-1-\emph{pathways} Alternatively Spliced}
\description{
The n1pas package provides functions to run N1PAS:
}
\section{N-of-1-\emph{pathways} functions}{

\itemize{
\code{\link{compute_multi_nof1_pathways}}
\code{\link{compute_nof1_md}}
\code{\link{compute_nof1_pathways}}
\code{\link{compute_nof1_wilcoxon}}
\code{\link{compute_vert_md}}
\code{\link{compute_zscore}}
}
}

