% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n1pas_method.R
\name{transform_gene_pathway}
\alias{transform_gene_pathway}
\title{N-of-1-\emph{pathways} Mahalanobis distance (MD) Calculation}
\usage{
transform_gene_pathway(gene_dist, annot_file, desc_file = NULL,
  genes_range = c(15, 500), fdr_threshold = 0.2, plot_locfdr = 0,
  small_ontology = TRUE, custom_locfdr = FALSE, ...)
}
\arguments{
\item{...}{Other arguments for stats::t.test.}

\item{baseline}{A vector (paired with case) containing all gene expression}

\item{case}{A vector (paired with baseline) containing all gene expression}
}
\value{
A list with the pathway_score (average MD score),
direction (up or down relative to the baseline),
and p_value. P_value is computed via a t.test.
}
\description{
\code{compute_nof1_md} computes N-of-1-\emph{pathways} Mahalanobis distance (MD)
statistic, direction, and P-value for a pathway (Schissler et al. 2015).
}
\details{
This function enables gene set (pathway) testing for a single pathway while providing a context-meaningful effect size. In the original publication by Schissler et al. 2015, the P-value was produced via a nonparametric bootstrapping procedure. However, in later yet unpublished studies, it has been determined that a simple paired \emph{t}-test performs as well and much faster. Therefore the \emph{t}-test has been implemented. For a nonparametric approach use N-of-1-\emph{pathways} Wilcoxon for Gardeux et al. 2014.
}
\examples{
## compute N-of-1-pathways MD for a single pathway
set.seed(44)
N <- 10
gene1 <- stats::rchisq(N, 5)
gene2 <- stats::rchisq(N, 5)
md_res <- compute_nof1_md(gene1, gene2)
print(md_res)

}
\references{
Gardeux, Vincent, et al. "'N-of-1-\emph{pathways}' unveils personal deregulated mechanisms from a single pair of RNA-Seq samples: towards precision medicine." Journal of the American Medical Informatics Association 21.6 (2014): 1015-1025.

Schissler, A. Grant, et al. "Dynamic changes of RNA-sequencing expression for precision medicine: N-of-1-\emph{pathways} Mahalanobis distance within pathways of single subjects predicts breast cancer survival." Bioinformatics 31.12 (2015): i293-i302.
}
\seealso{
\code{link{compute_nof1_wilcoxon}}, \code{link{compute_nof1_pathways}}, \code{link{compute_multi_nof1_pathways}}
}
