% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n1pas_method.R
\name{compute_n1pas}
\alias{compute_n1pas}
\title{Execute N-of-1-\emph{pathways} Alternatively Spliced (N1PAS)}
\usage{
compute_n1pas(iso_data, annot_file, desc_file = NULL, iso_range = c(2,
  30), genes_range = c(15, 500), ...)
}
\arguments{
\item{iso_data}{A data frame containing gene symbols, 'case' and 'baseline' isoform measurements. Each row corresponds to a unique isoform.}

\item{annot_file}{the path and file name for the gene set annotations}

\item{desc_file}{optionally provide the path and file name for gene set descriptions}

\item{iso_range}{a two-component vector to filter genes based on minimum and maximum number of isoforms. Default is (2, 30).}

\item{genes_range}{a two-component vector to filter gene sets based on minimum and maximum number of genes. Default is (15, 500).}

\item{...}{Other arguments for other n1pas functions}
}
\value{
pathway (gene set) enrichment profile organized in a data frame with 7 variables sorted by local fdr value.
\describe{
\item{row.names}{rows are labeled by user-specified gene set ID, the "path_id"}
\item{pathway_score}{numeric. N1PAS's effect size: an odds ratio quantifying pathawy enrichment of alternatively spliced genes}
\item{fdr_value}{numeric. local false discovery rate. See Efron 2013}
\item{num_genes_annot}{numeric. Number of genes annotated to the pathway from gene set input}
\item{num_genes_measured}{numeric. Number of genes measured and expressed (in at least one sample) within the pathway.}
\item{upper_fdr_threshold}{numeric. Odds ratios larger than this threshold are identified as enriched with alternatively spliced genes.}
\item{diff_splice_call}{numeric. Binary values with 1 being pathway is enriched and 0 otherwise.}
\item{pathway_desc}{character. Description of the pathway provided by the user.}
}
}
\description{
\code{compute_n1pas} transforms a pair of isoform expression data sets
to a profile of pathways enriched with alternatively spliced genes.
}
\details{
This function performs the entire N1PAS workflow for a pair of samples.
}
\note{
This function requires does not require library-size normalized data as the Hellinger distance is invariant under scaling.
}
\references{
Efron, B. (2013). “Local False Discovery Rates,” in Large-Scale Inference doi:10.1017/cbo9780511761362.006

Schissler, A. Grant, et al. "A single-subject method to detect pathway enriched with alternatively spliced genes." Frontiers in Genetics, to appear, (2019)

Gardeux, Vincent, et al. "'N-of-1-\emph{pathways}' unveils personal deregulated mechanisms from a single pair of RNA-Seq samples: towards precision medicine." Journal of the American Medical Informatics Association 21.6 (2014): 1015-1025.

Schissler, A. Grant, et al. "Dynamic changes of RNA-sequencing expression for precision medicine: N-of-1-\emph{pathways} Mahalanobis distance within pathways of single subjects predicts breast cancer survival." Bioinformatics 31.12 (2015): i293-i302.
}
\seealso{
\code{\link{compute_hellinger}} \code{\link{get_OR}} \code{link{transform_iso_gene}} \code{link{transform_gene_pathway}} \code{\link{blca_patient_iso_kegg}}
}
