% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n1pas_method.R
\name{get_OR}
\alias{get_OR}
\title{Compute Hellinger distance for gene-specific isoforms given pair of measurements}
\usage{
get_OR(tmp_genes, dist_data, genes_range = c(15, 500))
}
\arguments{
\item{X}{a N x matrix of isoform counts, values restricted between 0 and}

\item{direction}{a numeric vector indicating the direction of differentially expression
with 1 indicating higher expressed and -1 indicating lower expressed in the 'case' sample.}
}
\value{
a numeric vector of signed z-scores
}
\description{
\code{compute_hellinger} transforms a p_value vector to signed z-score vector.
}
\details{
This function provides the Hellinger distance between two samples.
This distance is used to quantify differential isoform usage.
PROVIDE FORMULA
}
\examples{
set.seed(44)
N <- 100
p_values <- stats::runif(N)
directions <- sample(c(-1,1), size = N, replace = TRUE)
zscores <- compute_zscore(p_values, directions)
\dontrun{
stats::qqnorm(zscores); abline(0,1)
}

}
