% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n1pas.R
\docType{package}
\name{n1pas}
\alias{n1pas}
\alias{n1pas-package}
\title{\code{n1pas}: An R Package for N-of-1-\emph{pathways} Alternatively Spliced}
\description{
The n1pas package provides functions to run N1PAS:
}
\details{
N1PAS functions:
\code{\link{compute_hellinger}}
\code{\link{get_OR}}
\code{\link{transform_iso_gene}}
\code{\link{transform_gene_pathway}}
\code{\link{compute_n1pas}}
}
